refactor(op-adapter): split lib.rs into per-plane modules
`crates/op-adapter/src/lib.rs` was 1454 lines — four plane encoders, the
credential predicates, the stand-in account seam, and ~900 lines of tests in
one `mod tests`. Same treatment as #3473 and #3474.

One module per plane, because the planes are what the coverage docs are
written against: `data`, `acl`, `group`, `root`, plus `credential` (the
op-local admission predicates the governance apply path shares) and `legacy`
(the stand-in account derivation, the part deleted first at cutover).
`lib.rs` keeps the crate docs and re-exports every public item flat, so
`calimero_op_adapter::payload_from_root_op` resolves exactly as before.

Tests move to `src/tests/<module>.rs`, one file per module under test, with
fixtures in `src/tests/support.rs`. All 12 tests keep their names, bodies and
assertions — `tests::root::root_op_encoder_mapping` is the old
`tests::root_op_encoder_mapping`.

No behavior change. Two things the split made visible and worth fixing:

- `calimero-authz` was a normal dependency used only by a test, so it moves to
  `[dev-dependencies]`.
- `AGENTS.md` documented `payload_from_root_op(op, signer)` and "6 unit
  tests"; the signer argument is gone and there are 12. Also documents the
  three `legacy` functions and the two credential predicates, which the
  Public API table never listed.

Verified: `cargo test -p calimero-op-adapter` 12/12 (same set as before the
split), `cargo fmt --check`, `cargo clippy --workspace --all-targets
--features calimero-storage/testing -- -D warnings`, `cargo check --workspace`,
`cargo doc -p calimero-op-adapter` all clean.
