+++ crates/op-adapter/Cargo.toml
@@ -9,7 +9,6 @@ license.workspace = true
 publish = false
 
 [dependencies]
-calimero-authz.workspace = true
 calimero-context-config.workspace = true
 calimero-governance-types.workspace = true
 sha2.workspace = true
@@ -20,5 +19,6 @@ calimero-primitives.workspace = true
 calimero-storage.workspace = true
 
 [dev-dependencies]
+calimero-authz.workspace = true
 calimero-primitives.workspace = true
 calimero-projection.workspace = true
